backend/requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
Automat==0.7.0
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
dialogflow==0.5.2
docutils==0.14
dpath==1.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
gast==0.2.2
gensim==3.7.2
google-api-core==1.8.1
google-auth==1.6.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.9.0
h5py==2.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
json-spec==0.10.1
jsonschema==2.6.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lxml==4.3.2
Markdown==3.1
MarkupSafe==1.1.1
mock==2.0.0
nltk==3.4
numpy==1.16.2
parsel==1.5.1
pbr==5.1.3
protobuf==3.7.0
py-stackexchange==2.2.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
pyOpenSSL==19.0.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==4.2b1
queuelib==1.5.0
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
scipy==1.2.1
Scrapy==1.6.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.1
swagger-py-codegen==0.4.0
swagger-spec-validator==2.4.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.20.0
Werkzeug==0.14.1
zope.interface==4.6.0
